// src/app/core/api/model/milestone-dto.ts
/**
 * Wedding API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MilestoneDtoTitle } from './milestone-dto-title';
import { MilestoneDtoAnnouncement } from './milestone-dto-announcement';


export interface MilestoneDto { 
    id: string;
    title: MilestoneDtoTitle;
    plannedDate: string;
    kind: MilestoneDto.KindEnum;
    reached: boolean;
    announcementType?: MilestoneDto.AnnouncementTypeEnum;
    audience?: MilestoneDto.AudienceEnum;
    announcement?: MilestoneDtoAnnouncement;
    version: number;
    atRisk: boolean;
}
export namespace MilestoneDto {
    export const KindEnum = {
        INTERNAL: 'internal',
        GUEST_FACING: 'guest-facing'
    } as const;
    export type KindEnum = typeof KindEnum[keyof typeof KindEnum];
    export const AnnouncementTypeEnum = {
        RSVP_REMINDER: 'rsvp-reminder',
        MENU_SELECTION_REMINDER: 'menu-selection-reminder'
    } as const;
    export type AnnouncementTypeEnum = typeof AnnouncementTypeEnum[keyof typeof AnnouncementTypeEnum];
    export const AudienceEnum = {
        ALL: 'all',
        NOT_REPLIED: 'not-replied',
        ATTENDING: 'attending',
        ATTENDING_NO_MENU: 'attending-no-menu'
    } as const;
    export type AudienceEnum = typeof AudienceEnum[keyof typeof AudienceEnum];
}