// src/app/core/api/model/notification-dto.ts
/**
 * Wedding API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface NotificationDto { 
    id: string;
    createdAt: string;
    type: string;
    templateId: string;
    title?: string;
    body?: string;
    data?: { [key: string]: any; };
    status: NotificationDto.StatusEnum;
    readAt?: string;
}
export namespace NotificationDto {
    export const StatusEnum = {
        UNREAD: 'unread',
        READ: 'read'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}